:sparkles: Add EditClockings function and refactor related structs
- Introduced `EditClockings` function in `rhApi.go` to handle manual clocking adjustments with error handling for date and hour parsing.
- Updated `RequestClockingBody` and `ResponseGetTime` structs to use `int64` for hour representation, improving compatibility with time calculations.
- Added new `EditClockingsBody` and `RequestEditClockingBody` structs in `structs.go` for better organization of edit clocking requests.
- Refactored `FormatTime` utility function to accept `int64` for current time, aligning with the updated struct definitions and enhancing time formatting capabilities.
- Improved overall code clarity and maintainability through consistent struct usage and error management.

+++ utils/main.go
@@ -4,9 +4,7 @@ import (
 	"time"
 )
 
-
-
-func FormatTime(currentDate string, currentTime uint32) (string, time.Time) {
+func FormatTime(currentDate string, currentTime int64) (string, time.Time) {
 	parsedDate, err := time.Parse(time.RFC3339, currentDate)
 	if err != nil {
 		return "", time.Time{}
@@ -14,14 +12,13 @@ func FormatTime(currentDate string, currentTime uint32) (string, time.Time) {
 
 	duration := time.Duration(currentTime) * time.Millisecond
 
-	newDate := time.Date(parsedDate.Year(), parsedDate.Month(), parsedDate.Day(), 0,0,0,0, time.Local).Add(duration)
+	newDate := time.Date(parsedDate.Year(), parsedDate.Month(), parsedDate.Day(), 0, 0, 0, 0, time.Local).Add(duration)
 	return newDate.Format("02/01/2006 15:04:05"), newDate
 }
-
 
 func FormatDuration(duration int32) string {
 
 	parsedDuration := time.Duration(duration) * time.Millisecond
 
-	return  parsedDuration.String()
+	return parsedDuration.String()
 }
